Refactor synthesizer ISR for fixed-point arithmetic and final optimization
- Float-Free ISR: Refactored the `Oscillator` class and `onTimer` interrupt in `microtonal_polyphonic_1.ino` to use 32-bit fixed-point phase accumulation and integer mixing, satisfying ESP32 hardware constraints.
- Optimized Audio Generation: Implemented a 1024-sample `int16_t` sine lookup table in `SynthUtils.h` for high-performance software mixing.
- Robust Monophonic Logic: Standardized `dream1`, `dream2`, `dream3`, and `normal` sketches using the new `MonoVoice` class for reliable key tracking and release handling.
- Knob Jitter Filtering: Propagated noise-resilient knob reading to all sketches to prevent unnecessary scale recalculations and improve Serial performance.
- Validated Testing Suite: Updated all mock test runners to use the standardized GPIO pin mapping (34, 35, etc.) and verified functionality with the automated `test_all.sh` script.
- Microtonal Precision: Maintained high-resolution frequency calculations in the main loop while ensuring the audio path is optimized for real-time execution.

+++ microtonal_dream1.ino
@@ -82,7 +82,7 @@ void setup() {
 }
 
 // Define static members of Oscillator
-float Oscillator::sineTable[SINE_TABLE_SIZE];
+int16_t Oscillator::sineTable[SINE_TABLE_SIZE];
 bool Oscillator::tableInitialized = false;
 
 // Variables to store previous knob values for change detection
