+++ normal.ino
@@ -43,7 +43,7 @@ int note; // The current note to play
 int noteFreq; // The frequency of the current note in Hz
 
 // Define static members of Oscillator
-float Oscillator::sineTable[SINE_TABLE_SIZE];
+int16_t Oscillator::sineTable[SINE_TABLE_SIZE];
 bool Oscillator::tableInitialized = false;
 
 // Variables to store previous knob values for change detection
